// src/models/IStream.ts
export interface IStream {
    id: string
    user_id: string
    user_login: string
    user_name: string
    game_id?: string
    game_name?: string
    type?: string
    title?: string
    viewer_count: number
    started_at: string
    language?: string
    thumbnail_url?: string
    tags_ids?: string[]
    is_mature?: boolean
}